chore: Update challenge configuration and host config
The challenge configuration file has been updated to change the title and short description of the challenge. Additionally, the remote evaluation and is_docker_based fields have been modified to enable remote evaluation and indicate that the challenge is docker-based.

The host config file has also been updated with a new authentication token, team PK, and EvalAI host URL.

These changes were made to align the challenge configuration and host config with the requirements of the project.